Franks Laboratory JavaScript Game Development
Adding animated sprite sheets to your JavaScript games is one of the fastest and the most impactful ways to bring your games to the next level and to add that professional look and polish you want. Games with sprites are great at pulling your audience into the fantasy you are trying to create and they are also really fun to make.